refactor: 패키지 경로 변경 및 변수명 변경
