+++ src/main/java/com/codecrafters/companity/adapter/post/dto/request/RequestForCreatingPost.java
@@ -1,7 +1,7 @@
 package com.codecrafters.companity.adapter.post.dto.request;
 
 import com.codecrafters.companity.adapter.post.mapper.PostMapperForController;
-import com.codecrafters.companity.application.in.post.dto.PostForCreate;
+import com.codecrafters.companity.domain.post.PostForCreate;
 import com.codecrafters.companity.domain.enums.City;
 import com.codecrafters.companity.domain.enums.Sport;
 import lombok.Getter;
